small

+++ labs/lab05/src/main/scala/example/Record.scala
@@ -56,5 +56,5 @@ case class Record (
   // Getters
   def getId: Int = id 
   def getState: String = state
-  def getSales: Int = sales
+  def getSales: Double = sales
 }
